requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
numpy==1.26.4
opencv-python==4.10.0.84
faiss-cpu==1.8.0.post1
redis==5.0.7
pydantic==2.8.2
jinja2==3.1.4
python-multipart==0.0.9